requirements.txt
together
anthropic
openai
typer
datasets
srsly
python-dotenv
pydantic
tqdm
pyyaml